// src/app/models/items/psychoMod.ts
export enum PsychoMod {
  CRIT_CHANCE = "CRIT_CHANCE",
  DOUBLE_HIT_CHANCE = "DOUBLE_HIT_CHANCE",

  CRIT_RESISTANCE = "CRIT_RESISTANCE",
  CRIT_DAMAGE_REDUCTION = "CRIT_DAMAGE_REDUCTION",

  MAGICAL_DAMAGE_INCREASE = "MAGICAL_DAMAGE_INCREASE",
  PHYSICAL_DAMAGE_INCREASE = "PHYSICAL_DAMAGE_INCREASE",

  EXTRA_AP = "EXTRA_AP",
  EXTRA_ATTACK_CIRCLE = "EXTRA_ATTACK_CIRCLE",
  MANA_DRAIN = "MANA_DRAIN",
  CHANCE_OF_DISENCHANTMENT = "CHANCE_OF_DISENCHANTMENT",

  MAGICAL_HIT_MODIFIER = "MAGICAL_HIT_MODIFIER",
  RANGE_HIT_MODIFIER = "RANGE_HIT_MODIFIER",
  PHYSICAL_HIT_MODIFIER = "PHYSICAL_HIT_MODIFIER",

  MAGICAL_DEFENCE_MODIFIER = "MAGICAL_DEFENCE_MODIFIER",
  RANGE_DEFENCE_MODIFIER = "RANGE_DEFENCE_MODIFIER",
  PHYSICAL_DEFENCE_MODIFIER = "PHYSICAL_DEFENCE_MODIFIER",

  EXTRA_FIRE_DAMAGE = "EXTRA_FIRE_DAMAGE",
  EXTRA_COLD_DAMAGE = "EXTRA_COLD_DAMAGE",
  EXTRA_ENERGY_DAMAGE = "EXTRA_ENERGY_DAMAGE",

  RESISTANCE_TO_ROOT = "RESISTANCE_TO_ROOT",
  RESISTANCE_TO_FREEZING = "RESISTANCE_TO_FREEZING",
  STUN_AND_INSTA_KILL_RESISTANCE = "STUN_AND_INSTA_KILL_RESISTANCE",

  MANA_REGENERATION = "MANA_REGENERATION",
  STAMINA_REGENERATION = "STAMINA_REGENERATION",
  MANA_USAGE = "MANA_USAGE",
  STAMINA_USAGE = "STAMINA_USAGE",

  FARID = "FARID",
  HOLM = "HOLM",
  PERCENTAGE_DAMAGE_REDUCTION = "PERCENTAGE_DAMAGE_REDUCTION",

  MENTAL_ATTACKS_PENETRATION = "MENTAL_ATTACKS_PENETRATION",

  DOUBLE_ATTACK_ROLL_CHANCE = "DOUBLE_ATTACK_ROLL_CHANCE",
  DOUBLE_DEFENCE_ROLL_CHANCE = "DOUBLE_DEFENCE_ROLL_CHANCE",

  PASSIVE_DAMAGE_REDUCTION = "PASSIVE_DAMAGE_REDUCTION",
  DAMAGE_REDUCTION = "DAMAGE_REDUCTION",

  CHEAT_DESTINY = "CHEAT_DESTINY"

}